Fade in and out
Fade animation for entering and leaving scenes added. There are two different animations, one for fade in and one for fade out. The animations simple raises or lowers the alpha value of a completely dark image. The image has been stolen from google and needs to be replaced in the future if it is supposed to be used.

Furthermore the pause canvas has been added to the game canvas. The way it works now is that you fill gameobjects onto an image and then you hide the image and all its children when you pause the game. The revamped pause script now uses two gameobjects, images, where it hides one and shows the other. On button press the shown image gets hidden and vice versa for the hidden image.

Lastly the scipt "LoadSceneOnSclick" has been revamped to work with the fading animation. It now uses an enumerator whereas the fading animaiton gets played both when entering and exiting a scenetransition.