Prefabs For Pause and Tutorial
Added prefabs for pause and tutorials.

The Pause prefab requires two kind of background images whereas one needs the entire main screen on top of it, whilst the other needs the entire pause menu on it. See PlaySceneTemp scene for an example on how it should work.

The tutorial prefab has had some changes to make it work as intended.